+++ setup.sql
@@ -68,7 +68,7 @@ CREATE TABLE recipe_ingredients (
     id SERIAL PRIMARY KEY
     , id_recipe INT REFERENCES recipes(id) NOT NULL
     , id_ingredient INT REFERENCES ingredients(id) NOT NULL
-    , quantity NUMERIC(10, 2) NOT NULL
+    , quantity INTEGER NOT NULL
     , id_unit INT REFERENCES units(id) NOT NULL
     , created_at TIMESTAMP DEFAULT NOW()
     , created_by VARCHAR(64) NOT NULL
